requirements/base.txt
# Core Framework
Django==5.2.6

# Environment and Configuration
django-environ==0.11.2

# Database and ORM
psycopg2-binary==2.9.9


# Security and Permissions
django-easy-audit==1.3.7
django-simple-history==3.10.1

# Middleware and Utilities
whitenoise==6.5.0
django-cors-headers==4.3.1
django-extensions==3.2.3

# Image Processing
pillow==10.4.0

# Filtering and Querying
django-filter==23.5

# File Storage
django-storages==1.14.2
django-minio-storage==0.5.8

# Asynchronous Support
channels==4.1.0
daphne==4.1.2

# Celery and Related Extensions
celery==5.5.3
django-celery-results==2.6.0
django-celery-beat==2.8.1

# Server and Deployment
gunicorn==22.0.0
setuptools==70.3.0

# Error Tracking
sentry-sdk==2.30.0

# GraphQL
strawberry-graphql==0.282.0
strawberry-graphql-django==0.65.1

# Authentication
strawberry-django-auth==0.378.4
